Extract Solaris common abstract bases and add JNA suffix
Rebase Solaris OSProcess, OSThread, FileSystem, NetworkParams, and
OperatingSystem onto abstract bases in oshi-common, with concrete
<Os>FooJNA (oshi-core) and <Os>FooFFM (oshi-core-ffm) subclasses, matching
the AIX and FreeBSD convention. OSFileStore has no native code, so it
becomes a single concrete common class like AixOSFileStore. The shared
attribute parsing now reads the common /proc carriers, removing the
duplicated updateAttributes logic between the JNA and FFM sides.

The JNA-only Kstat2 paths stay JNA-side: HAS_KSTAT2, the dual-path file
descriptor counts, system uptime/boot time, and the getSessions override
remain in SolarisOperatingSystemJNA / SolarisFileSystemJNA. FFM uses the
legacy kstat chain only, since kstat2 exists only on the JDK 17-capped
latest Solaris and illumos forked before it.

The driver split mirrors AIX: the buffer parsing lives in the common
PsInfo driver, the JNA PsInfo is renamed to PsInfoJNA and keeps only the
pread-based arg/env read, and PsInfoFFM keeps only the pargs-based read.
The now-unused psinfo/lwpsinfo/prusage JNA structs are removed from
SolarisLibc.

Also fixes an FFM gap: its prusage POJO had dropped pr_ioch, so FFM never
populated OSProcess.getBytesRead(); the common carrier restores it so both
implementations report it.
