+++ oshi-core-ffm/src/main/java/oshi/software/os/unix/solaris/SolarisFileSystemFFM.java
@@ -4,118 +4,19 @@
  */
 package oshi.software.os.unix.solaris;
 
-import java.io.File;
 import java.lang.foreign.MemorySegment;
-import java.nio.file.PathMatcher;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.ffm.util.platform.unix.solaris.KstatUtilFFM;
 import oshi.ffm.util.platform.unix.solaris.KstatUtilFFM.KstatChain;
-import oshi.software.common.AbstractFileSystem;
-import oshi.software.os.OSFileStore;
-import oshi.util.ExecutingCommand;
-import oshi.util.FileSystemUtil;
-import oshi.util.FileUtil;
-import oshi.util.ParseUtil;
+import oshi.software.common.os.unix.solaris.SolarisFileSystem;
 
+/**
+ * FFM-backed Solaris File System. Uses the legacy {@code kstat} chain only; Kstat2 exists only on the JDK 17-capped
+ * latest Solaris, so FFM (JDK 25) never needs it.
+ */
 @ThreadSafe
-public class SolarisFileSystemFFM extends AbstractFileSystem {
-
-    public static final String OSHI_SOLARIS_FS_PATH_EXCLUDES = "oshi.os.solaris.filesystem.path.excludes";
-    public static final String OSHI_SOLARIS_FS_PATH_INCLUDES = "oshi.os.solaris.filesystem.path.includes";
-    public static final String OSHI_SOLARIS_FS_VOLUME_EXCLUDES = "oshi.os.solaris.filesystem.volume.excludes";
-    public static final String OSHI_SOLARIS_FS_VOLUME_INCLUDES = "oshi.os.solaris.filesystem.volume.includes";
-
-    private static final List<PathMatcher> FS_PATH_EXCLUDES = FileSystemUtil
-            .loadAndParseFileSystemConfig(OSHI_SOLARIS_FS_PATH_EXCLUDES);
-    private static final List<PathMatcher> FS_PATH_INCLUDES = FileSystemUtil
-            .loadAndParseFileSystemConfig(OSHI_SOLARIS_FS_PATH_INCLUDES);
-    private static final List<PathMatcher> FS_VOLUME_EXCLUDES = FileSystemUtil
-            .loadAndParseFileSystemConfig(OSHI_SOLARIS_FS_VOLUME_EXCLUDES);
-    private static final List<PathMatcher> FS_VOLUME_INCLUDES = FileSystemUtil
-            .loadAndParseFileSystemConfig(OSHI_SOLARIS_FS_VOLUME_INCLUDES);
-
-    @Override
-    public List<OSFileStore> getFileStores(boolean localOnly) {
-        return getFileStoreMatching(null, localOnly);
-    }
-
-    static List<OSFileStore> getFileStoreMatching(String nameToMatch, boolean localOnly) {
-        List<OSFileStore> fsList = new ArrayList<>();
-
-        Map<String, Long> inodeFreeMap = new HashMap<>();
-        Map<String, Long> inodeTotalMap = new HashMap<>();
-        String key = null;
-        String total = null;
-        String free;
-        String command = "df -g" + (localOnly ? " -l" : "");
-        for (String line : ExecutingCommand.runNative(command)) {
-            if (line.startsWith("/")) {
-                key = ParseUtil.whitespaces.split(line)[0];
-                total = null;
-            } else if (line.contains("available") && line.contains("total files")) {
-                total = ParseUtil.getTextBetweenStrings(line, "available", "total files").trim();
-            } else if (line.contains("free files")) {
-                free = ParseUtil.getTextBetweenStrings(line, "", "free files").trim();
-                if (key != null && total != null) {
-                    inodeFreeMap.put(key, ParseUtil.parseLongOrDefault(free, 0L));
-                    inodeTotalMap.put(key, ParseUtil.parseLongOrDefault(total, 0L));
-                    key = null;
-                }
-            }
-        }
-
-        for (String fs : ExecutingCommand.runNative("cat /etc/mnttab")) {
-            String[] split = ParseUtil.whitespaces.split(fs);
-            if (split.length < 5) {
-                continue;
-            }
-            String volume = split[0];
-            String path = split[1];
-            String type = split[2];
-            String options = split[3];
-
-            boolean isLocal = !NETWORK_FS_TYPES.contains(type);
-            if ((localOnly && !isLocal)
-                    || !path.equals("/") && (PSEUDO_FS_TYPES.contains(type) || FileSystemUtil.isFileStoreExcluded(path,
-                            volume, FS_PATH_INCLUDES, FS_PATH_EXCLUDES, FS_VOLUME_INCLUDES, FS_VOLUME_EXCLUDES))) {
-                continue;
-            }
-
-            String name = path.substring(path.lastIndexOf('/') + 1);
-            if (name.isEmpty()) {
-                name = volume.substring(volume.lastIndexOf('/') + 1);
-            }
-
-            if (nameToMatch != null && !nameToMatch.equals(name)) {
-                continue;
-            }
-            File f = new File(path);
-            long totalSpace = f.getTotalSpace();
-            long usableSpace = f.getUsableSpace();
-            long freeSpace = f.getFreeSpace();
-
-            String description;
-            if (volume.startsWith("/dev") || path.equals("/")) {
-                description = "Local Disk";
-            } else if (volume.equals("tmpfs")) {
-                description = "Ram Disk";
-            } else if (NETWORK_FS_TYPES.contains(type)) {
-                description = "Network Disk";
-            } else {
-                description = "Mount Point";
-            }
-
-            fsList.add(new SolarisOSFileStoreFFM(name, volume, name, path, options, "", isLocal, "", description, type,
-                    freeSpace, usableSpace, totalSpace, inodeFreeMap.getOrDefault(path, 0L),
-                    inodeTotalMap.getOrDefault(path, 0L)));
-        }
-        return fsList;
-    }
+public final class SolarisFileSystemFFM extends SolarisFileSystem {
 
     @Override
     public long getOpenFileDescriptors() {
@@ -137,16 +38,5 @@ public class SolarisFileSystemFFM extends AbstractFileSystem {
             }
         }
         return 0L;
-    }
-
-    @Override
-    public long getMaxFileDescriptorsPerProcess() {
-        final List<String> lines = FileUtil.readFile("/etc/system");
-        for (final String line : lines) {
-            if (line.startsWith("set rlim_fd_max")) {
-                return ParseUtil.parseLastLong(line, 65536L);
-            }
-        }
-        return 65536L;
     }
 }
