+++ oshi-core/src/main/java/oshi/software/os/unix/solaris/SolarisOSThreadJNA.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2026 The OSHI Project Contributors
+ * Copyright 2020-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 package oshi.software.os.unix.solaris;
@@ -8,12 +8,12 @@ import oshi.annotation.concurrent.ThreadSafe;
 import oshi.software.common.os.unix.solaris.SolarisOSThread;
 
 /**
- * FFM-backed Solaris OSThread.
+ * JNA-backed Solaris OSThread.
  */
 @ThreadSafe
-public final class SolarisOSThreadFFM extends SolarisOSThread {
+public final class SolarisOSThreadJNA extends SolarisOSThread {
 
-    public SolarisOSThreadFFM(int pid, int lwpid) {
+    public SolarisOSThreadJNA(int pid, int lwpid) {
         super(pid, lwpid);
     }
 }
